Move Drawer to layouts
